refactor: 불필요한 파일 삭제 및 useMapDataUpdater 훅으로 비즈니스 로직 분리
